Struct aws_sdk_gamelift::client::fluent_builders::CreateMatchmakingConfiguration

pub struct CreateMatchmakingConfiguration<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

Implementations

impl<C, M, R> CreateMatchmakingConfiguration<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub async fn send(
self
) -> Result<CreateMatchmakingConfigurationOutput, SdkError<CreateMatchmakingConfigurationError>> where
R::Policy: SmithyRetryPolicy<CreateMatchmakingConfigurationInputOperationOutputAlias, CreateMatchmakingConfigurationOutput, CreateMatchmakingConfigurationError, CreateMatchmakingConfigurationInputOperationRetryAlias>,

pub fn name(self, inp: impl Into<String>) -> Self

A unique identifier for the matchmaking configuration. This name is used to identify the configuration associated with a matchmaking request or ticket.

pub fn set_name(self, input: Option<String>) -> Self

pub fn description(self, inp: impl Into<String>) -> Self

A human-readable description of the matchmaking configuration.

pub fn set_description(self, input: Option<String>) -> Self

pub fn game_session_queue_arns(self, inp: impl Into<String>) -> Self

Appends an item to GameSessionQueueArns.

To override the contents of this collection use set_game_session_queue_arns.

The Amazon Resource Name (ARN) that is assigned to a GameLift game session queue resource and uniquely identifies it. ARNs are unique across all Regions. Format is arn:aws:gamelift:::gamesessionqueue/. Queues can be located in any Region. Queues are used to start new GameLift-hosted game sessions for matches that are created with this matchmaking configuration. If FlexMatchMode is set to STANDALONE, do not set this parameter.

pub fn set_game_session_queue_arns(self, input: Option<Vec<String>>) -> Self

pub fn request_timeout_seconds(self, inp: i32) -> Self

The maximum duration, in seconds, that a matchmaking ticket can remain in process before timing out. Requests that fail due to timing out can be resubmitted as needed.

pub fn set_request_timeout_seconds(self, input: Option<i32>) -> Self

pub fn acceptance_timeout_seconds(self, inp: i32) -> Self

The length of time (in seconds) to wait for players to accept a proposed match, if acceptance is required.

pub fn set_acceptance_timeout_seconds(self, input: Option<i32>) -> Self

pub fn acceptance_required(self, inp: bool) -> Self

A flag that determines whether a match that was created with this configuration must be accepted by the matched players. To require acceptance, set to TRUE. With this option enabled, matchmaking tickets use the status REQUIRES_ACCEPTANCE to indicate when a completed potential match is waiting for player acceptance.

pub fn set_acceptance_required(self, input: Option<bool>) -> Self

pub fn rule_set_name(self, inp: impl Into<String>) -> Self

A unique identifier for the matchmaking rule set to use with this configuration. You can use either the rule set name or ARN value. A matchmaking configuration can only use rule sets that are defined in the same Region.

pub fn set_rule_set_name(self, input: Option<String>) -> Self

pub fn notification_target(self, inp: impl Into<String>) -> Self

An SNS topic ARN that is set up to receive matchmaking notifications. See Setting up notifications for matchmaking for more information.

pub fn set_notification_target(self, input: Option<String>) -> Self

pub fn additional_player_count(self, inp: i32) -> Self

The number of player slots in a match to keep open for future players. For example, if the configuration's rule set specifies a match for a single 12-person team, and the additional player count is set to 2, only 10 players are selected for the match. This parameter is not used if FlexMatchMode is set to STANDALONE.

pub fn set_additional_player_count(self, input: Option<i32>) -> Self

pub fn custom_event_data(self, inp: impl Into<String>) -> Self

Information to be added to all events related to this matchmaking configuration.

pub fn set_custom_event_data(self, input: Option<String>) -> Self

pub fn game_properties(self, inp: impl Into<GameProperty>) -> Self

Appends an item to GameProperties.

To override the contents of this collection use set_game_properties.

A set of custom properties for a game session, formatted as key:value pairs. These properties are passed to a game server process in the GameSession object with a request to start a new game session (see Start a Game Session). This information is added to the new GameSession object that is created for a successful match. This parameter is not used if FlexMatchMode is set to STANDALONE.

pub fn set_game_properties(self, input: Option<Vec<GameProperty>>) -> Self

pub fn game_session_data(self, inp: impl Into<String>) -> Self

A set of custom game session properties, formatted as a single string value. This data is passed to a game server process in the GameSession object with a request to start a new game session (see Start a Game Session). This information is added to the new GameSession object that is created for a successful match. This parameter is not used if FlexMatchMode is set to STANDALONE.

pub fn set_game_session_data(self, input: Option<String>) -> Self

pub fn backfill_mode(self, inp: BackfillMode) -> Self

The method used to backfill game sessions that are created with this matchmaking configuration. Specify MANUAL when your game manages backfill requests manually or does not use the match backfill feature. Specify AUTOMATIC to have GameLift create a StartMatchBackfill request whenever a game session has one or more open slots. Learn more about manual and automatic backfill in Backfill Existing Games with FlexMatch. Automatic backfill is not available when FlexMatchMode is set to STANDALONE.

pub fn set_backfill_mode(self, input: Option<BackfillMode>) -> Self

pub fn flex_match_mode(self, inp: FlexMatchMode) -> Self

Indicates whether this matchmaking configuration is being used with GameLift hosting or as a standalone matchmaking solution.

STANDALONE - FlexMatch forms matches and returns match information, including players and team assignments, in a MatchmakingSucceeded event.

WITH_QUEUE - FlexMatch forms matches and uses the specified GameLift queue to start a game session for the match.

pub fn set_flex_match_mode(self, input: Option<FlexMatchMode>) -> Self

pub fn tags(self, inp: impl Into<Tag>) -> Self

Appends an item to Tags.

To override the contents of this collection use set_tags.

A list of labels to assign to the new matchmaking configuration resource. Tags are developer-defined key-value pairs. Tagging AWS resources are useful for resource management, access management and cost allocation. For more information, see Tagging AWS Resources in the AWS General Reference. Once the resource is created, you can use TagResource, UntagResource, and ListTagsForResource to add, remove, and view tags. The maximum tag limit may be lower than stated. See the AWS General Reference for actual tagging limits.

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self