Struct aws_sdk_gamelift::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for GameLift.

This client allows ergonomic access to a GameLift-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

pub fn conf(&self) -> &Config

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn accept_match(&self) -> AcceptMatch<C, M, R>

pub fn claim_game_server(&self) -> ClaimGameServer<C, M, R>

pub fn create_alias(&self) -> CreateAlias<C, M, R>

pub fn create_build(&self) -> CreateBuild<C, M, R>

pub fn create_fleet(&self) -> CreateFleet<C, M, R>

pub fn create_fleet_locations(&self) -> CreateFleetLocations<C, M, R>

pub fn create_game_server_group(&self) -> CreateGameServerGroup<C, M, R>

pub fn create_game_session(&self) -> CreateGameSession<C, M, R>

pub fn create_game_session_queue(&self) -> CreateGameSessionQueue<C, M, R>

pub fn create_matchmaking_configuration(
&self
) -> CreateMatchmakingConfiguration<C, M, R>

pub fn create_matchmaking_rule_set(&self) -> CreateMatchmakingRuleSet<C, M, R>

pub fn create_player_session(&self) -> CreatePlayerSession<C, M, R>

pub fn create_player_sessions(&self) -> CreatePlayerSessions<C, M, R>

pub fn create_script(&self) -> CreateScript<C, M, R>

pub fn create_vpc_peering_authorization(
&self
) -> CreateVpcPeeringAuthorization<C, M, R>

pub fn create_vpc_peering_connection(
&self
) -> CreateVpcPeeringConnection<C, M, R>

pub fn delete_alias(&self) -> DeleteAlias<C, M, R>

pub fn delete_build(&self) -> DeleteBuild<C, M, R>

pub fn delete_fleet(&self) -> DeleteFleet<C, M, R>

pub fn delete_fleet_locations(&self) -> DeleteFleetLocations<C, M, R>

pub fn delete_game_server_group(&self) -> DeleteGameServerGroup<C, M, R>

pub fn delete_game_session_queue(&self) -> DeleteGameSessionQueue<C, M, R>

pub fn delete_matchmaking_configuration(
&self
) -> DeleteMatchmakingConfiguration<C, M, R>

pub fn delete_matchmaking_rule_set(&self) -> DeleteMatchmakingRuleSet<C, M, R>

pub fn delete_scaling_policy(&self) -> DeleteScalingPolicy<C, M, R>

pub fn delete_script(&self) -> DeleteScript<C, M, R>

pub fn delete_vpc_peering_authorization(
&self
) -> DeleteVpcPeeringAuthorization<C, M, R>

pub fn delete_vpc_peering_connection(
&self
) -> DeleteVpcPeeringConnection<C, M, R>

pub fn deregister_game_server(&self) -> DeregisterGameServer<C, M, R>

pub fn describe_alias(&self) -> DescribeAlias<C, M, R>

pub fn describe_build(&self) -> DescribeBuild<C, M, R>

pub fn describe_ec2_instance_limits(&self) -> DescribeEC2InstanceLimits<C, M, R>

pub fn describe_fleet_attributes(&self) -> DescribeFleetAttributes<C, M, R>

pub fn describe_fleet_capacity(&self) -> DescribeFleetCapacity<C, M, R>

pub fn describe_fleet_events(&self) -> DescribeFleetEvents<C, M, R>

pub fn describe_fleet_location_attributes(
&self
) -> DescribeFleetLocationAttributes<C, M, R>

pub fn describe_fleet_location_capacity(
&self
) -> DescribeFleetLocationCapacity<C, M, R>

pub fn describe_fleet_location_utilization(
&self
) -> DescribeFleetLocationUtilization<C, M, R>

pub fn describe_fleet_port_settings(&self) -> DescribeFleetPortSettings<C, M, R>

pub fn describe_fleet_utilization(&self) -> DescribeFleetUtilization<C, M, R>

pub fn describe_game_server(&self) -> DescribeGameServer<C, M, R>

pub fn describe_game_server_group(&self) -> DescribeGameServerGroup<C, M, R>

pub fn describe_game_server_instances(
&self
) -> DescribeGameServerInstances<C, M, R>

pub fn describe_game_session_details(
&self
) -> DescribeGameSessionDetails<C, M, R>

pub fn describe_game_session_placement(
&self
) -> DescribeGameSessionPlacement<C, M, R>

pub fn describe_game_session_queues(&self) -> DescribeGameSessionQueues<C, M, R>

pub fn describe_game_sessions(&self) -> DescribeGameSessions<C, M, R>

pub fn describe_instances(&self) -> DescribeInstances<C, M, R>

pub fn describe_matchmaking(&self) -> DescribeMatchmaking<C, M, R>

pub fn describe_matchmaking_configurations(
&self
) -> DescribeMatchmakingConfigurations<C, M, R>

pub fn describe_matchmaking_rule_sets(
&self
) -> DescribeMatchmakingRuleSets<C, M, R>

pub fn describe_player_sessions(&self) -> DescribePlayerSessions<C, M, R>

pub fn describe_runtime_configuration(
&self
) -> DescribeRuntimeConfiguration<C, M, R>

pub fn describe_scaling_policies(&self) -> DescribeScalingPolicies<C, M, R>

pub fn describe_script(&self) -> DescribeScript<C, M, R>

pub fn describe_vpc_peering_authorizations(
&self
) -> DescribeVpcPeeringAuthorizations<C, M, R>

pub fn describe_vpc_peering_connections(
&self
) -> DescribeVpcPeeringConnections<C, M, R>

pub fn get_game_session_log_url(&self) -> GetGameSessionLogUrl<C, M, R>

pub fn get_instance_access(&self) -> GetInstanceAccess<C, M, R>

pub fn list_aliases(&self) -> ListAliases<C, M, R>

pub fn list_builds(&self) -> ListBuilds<C, M, R>

pub fn list_fleets(&self) -> ListFleets<C, M, R>

pub fn list_game_server_groups(&self) -> ListGameServerGroups<C, M, R>

pub fn list_game_servers(&self) -> ListGameServers<C, M, R>

pub fn list_scripts(&self) -> ListScripts<C, M, R>

pub fn list_tags_for_resource(&self) -> ListTagsForResource<C, M, R>

pub fn put_scaling_policy(&self) -> PutScalingPolicy<C, M, R>

pub fn register_game_server(&self) -> RegisterGameServer<C, M, R>

pub fn request_upload_credentials(&self) -> RequestUploadCredentials<C, M, R>

pub fn resolve_alias(&self) -> ResolveAlias<C, M, R>

pub fn resume_game_server_group(&self) -> ResumeGameServerGroup<C, M, R>

pub fn search_game_sessions(&self) -> SearchGameSessions<C, M, R>

pub fn start_fleet_actions(&self) -> StartFleetActions<C, M, R>

pub fn start_game_session_placement(&self) -> StartGameSessionPlacement<C, M, R>

pub fn start_match_backfill(&self) -> StartMatchBackfill<C, M, R>

pub fn start_matchmaking(&self) -> StartMatchmaking<C, M, R>

pub fn stop_fleet_actions(&self) -> StopFleetActions<C, M, R>

pub fn stop_game_session_placement(&self) -> StopGameSessionPlacement<C, M, R>

pub fn stop_matchmaking(&self) -> StopMatchmaking<C, M, R>

pub fn suspend_game_server_group(&self) -> SuspendGameServerGroup<C, M, R>

pub fn tag_resource(&self) -> TagResource<C, M, R>

pub fn untag_resource(&self) -> UntagResource<C, M, R>

pub fn update_alias(&self) -> UpdateAlias<C, M, R>

pub fn update_build(&self) -> UpdateBuild<C, M, R>

pub fn update_fleet_attributes(&self) -> UpdateFleetAttributes<C, M, R>

pub fn update_fleet_capacity(&self) -> UpdateFleetCapacity<C, M, R>

pub fn update_fleet_port_settings(&self) -> UpdateFleetPortSettings<C, M, R>

pub fn update_game_server(&self) -> UpdateGameServer<C, M, R>

pub fn update_game_server_group(&self) -> UpdateGameServerGroup<C, M, R>

pub fn update_game_session(&self) -> UpdateGameSession<C, M, R>

pub fn update_game_session_queue(&self) -> UpdateGameSessionQueue<C, M, R>

pub fn update_matchmaking_configuration(
&self
) -> UpdateMatchmakingConfiguration<C, M, R>

pub fn update_runtime_configuration(
&self
) -> UpdateRuntimeConfiguration<C, M, R>

pub fn update_script(&self) -> UpdateScript<C, M, R>

pub fn validate_matchmaking_rule_set(
&self
) -> ValidateMatchmakingRuleSet<C, M, R>

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

pub fn from_conf(conf: Config) -> Self