Module aws_sdk_gamelift::error

Modules

conflict_exception

See ConflictException

fleet_capacity_exceeded_exception

See FleetCapacityExceededException

game_session_full_exception

See GameSessionFullException

idempotent_parameter_mismatch_exception

See IdempotentParameterMismatchException

internal_service_exception

See InternalServiceException

invalid_fleet_status_exception

See InvalidFleetStatusException

invalid_game_session_status_exception

See InvalidGameSessionStatusException

invalid_request_exception

See InvalidRequestException

limit_exceeded_exception

See LimitExceededException

not_found_exception

See NotFoundException

out_of_capacity_exception

See OutOfCapacityException

tagging_failed_exception

See TaggingFailedException

terminal_routing_strategy_exception

See TerminalRoutingStrategyException

unauthorized_exception

See UnauthorizedException

unsupported_region_exception

See UnsupportedRegionException

Structs

AcceptMatchError

ClaimGameServerError

ConflictException

The requested operation would cause a conflict with the current state of a service

CreateAliasError

CreateBuildError

CreateFleetError

CreateFleetLocationsError

CreateGameServerGroupError

CreateGameSessionError

CreateGameSessionQueueError

CreateMatchmakingConfigurationError

CreateMatchmakingRuleSetError

CreatePlayerSessionError

CreatePlayerSessionsError

CreateScriptError

CreateVpcPeeringAuthorizationError

CreateVpcPeeringConnectionError

DeleteAliasError

DeleteBuildError

DeleteFleetError

DeleteFleetLocationsError

DeleteGameServerGroupError

DeleteGameSessionQueueError

DeleteMatchmakingConfigurationError

DeleteMatchmakingRuleSetError

DeleteScalingPolicyError

DeleteScriptError

DeleteVpcPeeringAuthorizationError

DeleteVpcPeeringConnectionError

DeregisterGameServerError

DescribeAliasError

DescribeBuildError

DescribeEC2InstanceLimitsError

DescribeFleetAttributesError

DescribeFleetCapacityError

DescribeFleetEventsError

DescribeFleetLocationAttributesError

DescribeFleetLocationCapacityError

DescribeFleetLocationUtilizationError

DescribeFleetPortSettingsError

DescribeFleetUtilizationError

DescribeGameServerError

DescribeGameServerGroupError

DescribeGameServerInstancesError

DescribeGameSessionDetailsError

DescribeGameSessionPlacementError

DescribeGameSessionQueuesError

DescribeGameSessionsError

DescribeInstancesError

DescribeMatchmakingConfigurationsError

DescribeMatchmakingError

DescribeMatchmakingRuleSetsError

DescribePlayerSessionsError

DescribeRuntimeConfigurationError

DescribeScalingPoliciesError

DescribeScriptError

DescribeVpcPeeringAuthorizationsError

DescribeVpcPeeringConnectionsError

FleetCapacityExceededException

The specified fleet has no available instances to fulfill a

GameSessionFullException

The game instance is currently full and cannot allow the requested player(s) to

GetGameSessionLogUrlError

GetInstanceAccessError

IdempotentParameterMismatchException

A game session with this custom ID string already exists in this fleet. Resolve

InternalServiceException

The service encountered an unrecoverable internal failure while processing the

InvalidFleetStatusException

The requested operation would cause a conflict with the current state of a resource

InvalidGameSessionStatusException

The requested operation would cause a conflict with the current state of a resource

InvalidRequestException

One or more parameter values in the request are invalid. Correct the invalid

LimitExceededException

The requested operation would cause the resource to exceed the allowed service

ListAliasesError

ListBuildsError

ListFleetsError

ListGameServerGroupsError

ListGameServersError

ListScriptsError

ListTagsForResourceError

NotFoundException

A service resource associated with the request could not be found. Clients should

OutOfCapacityException

The specified game server group has no available game servers to fulfill a

PutScalingPolicyError

RegisterGameServerError

RequestUploadCredentialsError

ResolveAliasError

ResumeGameServerGroupError

SearchGameSessionsError

StartFleetActionsError

StartGameSessionPlacementError

StartMatchBackfillError

StartMatchmakingError

StopFleetActionsError

StopGameSessionPlacementError

StopMatchmakingError

SuspendGameServerGroupError

TagResourceError

TaggingFailedException

TerminalRoutingStrategyException

The service is unable to resolve the routing for a particular alias because it has

UnauthorizedException

The client failed authentication. Clients should not retry such requests.

UnsupportedRegionException

The requested operation is not supported in the Region specified.

UntagResourceError

UpdateAliasError

UpdateBuildError

UpdateFleetAttributesError

UpdateFleetCapacityError

UpdateFleetPortSettingsError

UpdateGameServerError

UpdateGameServerGroupError

UpdateGameSessionError

UpdateGameSessionQueueError

UpdateMatchmakingConfigurationError

UpdateRuntimeConfigurationError

UpdateScriptError

ValidateMatchmakingRuleSetError

Enums

AcceptMatchErrorKind

ClaimGameServerErrorKind

CreateAliasErrorKind

CreateBuildErrorKind

CreateFleetErrorKind

CreateFleetLocationsErrorKind

CreateGameServerGroupErrorKind

CreateGameSessionErrorKind

CreateGameSessionQueueErrorKind

CreateMatchmakingConfigurationErrorKind

CreateMatchmakingRuleSetErrorKind

CreatePlayerSessionErrorKind

CreatePlayerSessionsErrorKind

CreateScriptErrorKind

CreateVpcPeeringAuthorizationErrorKind

CreateVpcPeeringConnectionErrorKind

DeleteAliasErrorKind

DeleteBuildErrorKind

DeleteFleetErrorKind

DeleteFleetLocationsErrorKind

DeleteGameServerGroupErrorKind

DeleteGameSessionQueueErrorKind

DeleteMatchmakingConfigurationErrorKind

DeleteMatchmakingRuleSetErrorKind

DeleteScalingPolicyErrorKind

DeleteScriptErrorKind

DeleteVpcPeeringAuthorizationErrorKind

DeleteVpcPeeringConnectionErrorKind

DeregisterGameServerErrorKind

DescribeAliasErrorKind

DescribeBuildErrorKind

DescribeEC2InstanceLimitsErrorKind

DescribeFleetAttributesErrorKind

DescribeFleetCapacityErrorKind

DescribeFleetEventsErrorKind

DescribeFleetLocationAttributesErrorKind

DescribeFleetLocationCapacityErrorKind

DescribeFleetLocationUtilizationErrorKind

DescribeFleetPortSettingsErrorKind

DescribeFleetUtilizationErrorKind

DescribeGameServerErrorKind

DescribeGameServerGroupErrorKind

DescribeGameServerInstancesErrorKind

DescribeGameSessionDetailsErrorKind

DescribeGameSessionPlacementErrorKind

DescribeGameSessionQueuesErrorKind

DescribeGameSessionsErrorKind

DescribeInstancesErrorKind

DescribeMatchmakingConfigurationsErrorKind

DescribeMatchmakingErrorKind

DescribeMatchmakingRuleSetsErrorKind

DescribePlayerSessionsErrorKind

DescribeRuntimeConfigurationErrorKind

DescribeScalingPoliciesErrorKind

DescribeScriptErrorKind

DescribeVpcPeeringAuthorizationsErrorKind

DescribeVpcPeeringConnectionsErrorKind

GetGameSessionLogUrlErrorKind

GetInstanceAccessErrorKind

ListAliasesErrorKind

ListBuildsErrorKind

ListFleetsErrorKind

ListGameServerGroupsErrorKind

ListGameServersErrorKind

ListScriptsErrorKind

ListTagsForResourceErrorKind

PutScalingPolicyErrorKind

RegisterGameServerErrorKind

RequestUploadCredentialsErrorKind

ResolveAliasErrorKind

ResumeGameServerGroupErrorKind

SearchGameSessionsErrorKind

StartFleetActionsErrorKind

StartGameSessionPlacementErrorKind

StartMatchBackfillErrorKind

StartMatchmakingErrorKind

StopFleetActionsErrorKind

StopGameSessionPlacementErrorKind

StopMatchmakingErrorKind

SuspendGameServerGroupErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateAliasErrorKind

UpdateBuildErrorKind

UpdateFleetAttributesErrorKind

UpdateFleetCapacityErrorKind

UpdateFleetPortSettingsErrorKind

UpdateGameServerErrorKind

UpdateGameServerGroupErrorKind

UpdateGameSessionErrorKind

UpdateGameSessionQueueErrorKind

UpdateMatchmakingConfigurationErrorKind

UpdateRuntimeConfigurationErrorKind

UpdateScriptErrorKind

ValidateMatchmakingRuleSetErrorKind