Struct aws_sdk_gamelift::input::create_fleet_input::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for CreateFleetInput

Implementations

impl Builder

pub fn name(self, input: impl Into<String>) -> Self

A descriptive label that is associated with a fleet. Fleet names do not need to be unique.

pub fn set_name(self, input: Option<String>) -> Self

pub fn description(self, input: impl Into<String>) -> Self

A human-readable description of the fleet.

pub fn set_description(self, input: Option<String>) -> Self

pub fn build_id(self, input: impl Into<String>) -> Self

The unique identifier for a custom game server build to be deployed on fleet instances. You can use either the build ID or ARN. The build must be uploaded to GameLift and in READY status. This fleet property cannot be changed later.

pub fn set_build_id(self, input: Option<String>) -> Self

pub fn script_id(self, input: impl Into<String>) -> Self

The unique identifier for a Realtime configuration script to be deployed on fleet instances. You can use either the script ID or ARN. Scripts must be uploaded to GameLift prior to creating the fleet. This fleet property cannot be changed later.

pub fn set_script_id(self, input: Option<String>) -> Self

pub fn server_launch_path(self, input: impl Into<String>) -> Self

This parameter is no longer used. Specify a server launch path using the RuntimeConfiguration parameter. Requests that use this parameter instead continue to be valid.

pub fn set_server_launch_path(self, input: Option<String>) -> Self

pub fn server_launch_parameters(self, input: impl Into<String>) -> Self

This parameter is no longer used. Specify server launch parameters using the RuntimeConfiguration parameter. Requests that use this parameter instead continue to be valid.

pub fn set_server_launch_parameters(self, input: Option<String>) -> Self

pub fn log_paths(self, input: impl Into<String>) -> Self

pub fn set_log_paths(self, input: Option<Vec<String>>) -> Self

pub fn ec2_instance_type(self, input: Ec2InstanceType) -> Self

The GameLift-supported EC2 instance type to use for all fleet instances. Instance type determines the computing resources that will be used to host your game servers, including CPU, memory, storage, and networking capacity. See Amazon EC2 Instance Types for detailed descriptions of EC2 instance types.

pub fn set_ec2_instance_type(self, input: Option<Ec2InstanceType>) -> Self

pub fn ec2_inbound_permissions(self, input: impl Into<IpPermission>) -> Self

pub fn set_ec2_inbound_permissions(
self,
input: Option<Vec<IpPermission>>
) -> Self

pub fn new_game_session_protection_policy(self, input: ProtectionPolicy) -> Self

The status of termination protection for active game sessions on the fleet. By default, this property is set to NoProtection. You can also set game session protection for an individual game session by calling UpdateGameSession.

NoProtection - Game sessions can be terminated during active gameplay as a result of a scale-down event.

FullProtection - Game sessions in ACTIVE status cannot be terminated during a scale-down event.

pub fn set_new_game_session_protection_policy(
self,
input: Option<ProtectionPolicy>
) -> Self

pub fn runtime_configuration(self, input: RuntimeConfiguration) -> Self

Instructions for how to launch and maintain server processes on instances in the fleet. The runtime configuration defines one or more server process configurations, each identifying a build executable or Realtime script file and the number of processes of that type to run concurrently.

The RuntimeConfiguration parameter is required unless the fleet is being configured using the older parameters ServerLaunchPath and ServerLaunchParameters, which are still supported for backward compatibility.

pub fn set_runtime_configuration(
self,
input: Option<RuntimeConfiguration>
) -> Self

pub fn resource_creation_limit_policy(
self,
input: ResourceCreationLimitPolicy
) -> Self

A policy that limits the number of game sessions that an individual player can create on instances in this fleet within a specified span of time.

pub fn set_resource_creation_limit_policy(
self,
input: Option<ResourceCreationLimitPolicy>
) -> Self

pub fn metric_groups(self, input: impl Into<String>) -> Self

pub fn set_metric_groups(self, input: Option<Vec<String>>) -> Self

pub fn peer_vpc_aws_account_id(self, input: impl Into<String>) -> Self

Used when peering your GameLift fleet with a VPC, the unique identifier for the AWS account that owns the VPC. You can find your account ID in the AWS Management Console under account settings.

pub fn set_peer_vpc_aws_account_id(self, input: Option<String>) -> Self

pub fn peer_vpc_id(self, input: impl Into<String>) -> Self

A unique identifier for a VPC with resources to be accessed by your GameLift fleet. The VPC must be in the same Region as your fleet. To look up a VPC ID, use the VPC Dashboard in the AWS Management Console. Learn more about VPC peering in VPC Peering with GameLift Fleets.

pub fn set_peer_vpc_id(self, input: Option<String>) -> Self

pub fn fleet_type(self, input: FleetType) -> Self

Indicates whether to use On-Demand or Spot instances for this fleet. By default, this property is set to ON_DEMAND. Learn more about when to use On-Demand versus Spot Instances. This property cannot be changed after the fleet is created.

pub fn set_fleet_type(self, input: Option<FleetType>) -> Self

pub fn instance_role_arn(self, input: impl Into<String>) -> Self

A unique identifier for an AWS IAM role that manages access to your AWS services. With an instance role ARN set, any application that runs on an instance in this fleet can assume the role, including install scripts, server processes, and daemons (background processes). Create a role or look up a role's ARN by using the IAM dashboard in the AWS Management Console. Learn more about using on-box credentials for your game servers at Access external resources from a game server. This property cannot be changed after the fleet is created.

pub fn set_instance_role_arn(self, input: Option<String>) -> Self

pub fn certificate_configuration(self, input: CertificateConfiguration) -> Self

Prompts GameLift to generate a TLS/SSL certificate for the fleet. TLS certificates are used for encrypting traffic between game clients and the game servers that are running on GameLift. By default, the CertificateConfiguration is set to DISABLED. Learn more at Securing Client/Server Communication. This property cannot be changed after the fleet is created.

Note: This feature requires the AWS Certificate Manager (ACM) service, which is not available in all AWS regions. When working in a region that does not support this feature, a fleet creation request with certificate generation fails with a 4xx error.

pub fn set_certificate_configuration(
self,
input: Option<CertificateConfiguration>
) -> Self

pub fn locations(self, input: impl Into<LocationConfiguration>) -> Self

pub fn set_locations(self, input: Option<Vec<LocationConfiguration>>) -> Self

pub fn tags(self, input: impl Into<Tag>) -> Self

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self

pub fn build(self) -> Result<CreateFleetInput, BuildError>

Consumes the builder and constructs a CreateFleetInput