Module aws_sdk_gamelift::model

Modules

alias

See Alias

attribute_value

See AttributeValue

aws_credentials

See AwsCredentials

build

See Build

certificate_configuration

See CertificateConfiguration

desired_player_session

See DesiredPlayerSession

ec2_instance_counts

See Ec2InstanceCounts

ec2_instance_limit

See Ec2InstanceLimit

event

See Event

filter_configuration

See FilterConfiguration

fleet_attributes

See FleetAttributes

fleet_capacity

See FleetCapacity

fleet_utilization

See FleetUtilization

game_property

See GameProperty

game_server

See GameServer

game_server_group

See GameServerGroup

game_server_group_auto_scaling_policy

See GameServerGroupAutoScalingPolicy

game_server_instance

See GameServerInstance

game_session

See GameSession

game_session_connection_info

See GameSessionConnectionInfo

game_session_detail

See GameSessionDetail

game_session_placement

See GameSessionPlacement

game_session_queue

See GameSessionQueue

game_session_queue_destination

See GameSessionQueueDestination

instance

See Instance

instance_access

See InstanceAccess

instance_credentials

See InstanceCredentials

instance_definition

See InstanceDefinition

ip_permission

See IpPermission

launch_template_specification

See LaunchTemplateSpecification

location_attributes

See LocationAttributes

location_configuration

See LocationConfiguration

location_state

See LocationState

matched_player_session

See MatchedPlayerSession

matchmaking_configuration

See MatchmakingConfiguration

matchmaking_rule_set

See MatchmakingRuleSet

matchmaking_ticket

See MatchmakingTicket

placed_player_session

See PlacedPlayerSession

player

See Player

player_latency

See PlayerLatency

player_latency_policy

See PlayerLatencyPolicy

player_session

See PlayerSession

priority_configuration

See PriorityConfiguration

resource_creation_limit_policy

See ResourceCreationLimitPolicy

routing_strategy

See RoutingStrategy

runtime_configuration

See RuntimeConfiguration

s3_location

See S3Location

scaling_policy

See ScalingPolicy

script

See Script

server_process

See ServerProcess

tag

See Tag

target_configuration

See TargetConfiguration

target_tracking_configuration

See TargetTrackingConfiguration

vpc_peering_authorization

See VpcPeeringAuthorization

vpc_peering_connection

See VpcPeeringConnection

vpc_peering_connection_status

See VpcPeeringConnectionStatus

Structs

Alias

Properties that describe an alias resource.

AttributeValue

Values for use in Player attribute key-value pairs. This object lets

AwsCredentials

Temporary access credentials used for uploading game build files to Amazon GameLift. They

Build

Properties describing a custom game build.

CertificateConfiguration

Determines whether a TLS/SSL certificate is generated for a fleet. This feature must be

DesiredPlayerSession

Player information for use when creating player sessions using a game session

Ec2InstanceCounts

Resource capacity settings. Fleet capacity is measured in EC2 instances. Pending and

Ec2InstanceLimit

The GameLift service limits for an EC2 instance type and current utilization. GameLift

Event

Log entry describing an event that involves GameLift resources (such as a fleet). In

FilterConfiguration

A list of fleet locations where a game session queue can place new game sessions. You

FleetAttributes

Describes a GameLift fleet of game hosting resources.

FleetCapacity

Current resource capacity settings in a specified fleet or location. The location

FleetUtilization

Current resource utilization statistics in a specified fleet or location. The location

GameProperty

Set of key-value pairs that contain information about a game session. When included in

GameServer

GameServerGroup

GameServerGroupAutoScalingPolicy

GameServerInstance

GameSession

Properties describing a game session.

GameSessionConnectionInfo

Connection information for a new game session that is created in response to a StartMatchmaking request. Once a match is made, the FlexMatch engine

GameSessionDetail

A game session's properties plus the protection policy currently in

GameSessionPlacement

Object that describes a StartGameSessionPlacement request. This

GameSessionQueue

Configuration for a game session placement mechanism that processes requests for new

GameSessionQueueDestination

A fleet or alias designated in a game session queue. Queues fulfill requests for new

Instance

Represents an EC2 instance of virtual computing resources that hosts one or more game

InstanceAccess

Information required to remotely connect to a fleet instance. Access is requested

InstanceCredentials

Set of credentials required to remotely access a fleet instance. Access credentials

InstanceDefinition

IpPermission

A range of IP addresses and port settings that allow inbound traffic to connect to

LaunchTemplateSpecification

LocationAttributes

Represents a location in a multi-location fleet.

LocationConfiguration

A remote location where a multi-location fleet can deploy EC2 instances for game

LocationState

A fleet location and its life-cycle state. A location state object might

MatchedPlayerSession

Represents a new player session that is created as a result of a successful FlexMatch

MatchmakingConfiguration

Guidelines for use with FlexMatch to match players into games. All matchmaking

MatchmakingRuleSet

Set of rule statements, used with FlexMatch, that determine how to build your player

MatchmakingTicket

Ticket generated to track the progress of a matchmaking request. Each ticket is

PlacedPlayerSession

Information about a player session that was created as part of a StartGameSessionPlacement request. This object contains only the player

Player

Represents a player in matchmaking. When starting a matchmaking request, a player

PlayerLatency

Regional latency information for a player, used when requesting a new game session

PlayerLatencyPolicy

Sets a latency cap for individual players when placing a game session. With a latency

PlayerSession

Represents a player session. Player sessions are created either for a specific game

PriorityConfiguration

Custom prioritization settings for use by a game session queue when placing new game

ResourceCreationLimitPolicy

A policy that puts limits on the number of game sessions that a player can create

RoutingStrategy

The routing configuration for a fleet alias.

RuntimeConfiguration

A collection of server process configurations that describe the set of processes to

S3Location

The location in Amazon S3 where build or script files are stored for access by Amazon GameLift. This

ScalingPolicy

Rule that controls how a fleet is scaled. Scaling policies are uniquely identified

Script

Properties describing a Realtime script.

ServerProcess

A set of instructions for launching server processes on each instance in a fleet.

Tag

TargetConfiguration

Settings for a target-based scaling policy (see ScalingPolicy. A

TargetTrackingConfiguration

VpcPeeringAuthorization

Represents an authorization for a VPC peering connection between the VPC for an

VpcPeeringConnection

Represents a peering connection between a VPC on one of your AWS accounts and the

VpcPeeringConnectionStatus

Represents status information for a VPC peering connection. Status is associated

Enums

AcceptanceType

BackfillMode

BalancingStrategy

BuildStatus

CertificateType

ComparisonOperatorType

Ec2InstanceType

EventCode

FleetAction

FleetStatus

FleetType

FlexMatchMode

GameServerClaimStatus

GameServerGroupAction

GameServerGroupDeleteOption

GameServerGroupInstanceType

GameServerGroupStatus

GameServerHealthCheck

GameServerInstanceStatus

GameServerProtectionPolicy

GameServerUtilizationStatus

GameSessionPlacementState

GameSessionStatus

GameSessionStatusReason

InstanceStatus

IpProtocol

LocationUpdateStatus

MatchmakingConfigurationStatus

MetricName

OperatingSystem

PlayerSessionCreationPolicy

PlayerSessionStatus

PolicyType

PriorityType

ProtectionPolicy

RoutingStrategyType

ScalingAdjustmentType

ScalingStatusType

SortOrder