Struct aws_sdk_gamelift::model::game_session::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for GameSession

Implementations

impl Builder

pub fn game_session_id(self, input: impl Into<String>) -> Self

A unique identifier for the game session. A game session ARN has the following format: arn:aws:gamelift:::gamesession//.

pub fn set_game_session_id(self, input: Option<String>) -> Self

pub fn name(self, input: impl Into<String>) -> Self

A descriptive label that is associated with a game session. Session names do not need to be unique.

pub fn set_name(self, input: Option<String>) -> Self

pub fn fleet_id(self, input: impl Into<String>) -> Self

A unique identifier for the fleet that the game session is running on.

pub fn set_fleet_id(self, input: Option<String>) -> Self

pub fn fleet_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) associated with the GameLift fleet that this game session is running on.

pub fn set_fleet_arn(self, input: Option<String>) -> Self

pub fn creation_time(self, input: Instant) -> Self

A time stamp indicating when this data object was created. Format is a number expressed in Unix time as milliseconds (for example "1469498468.057").

pub fn set_creation_time(self, input: Option<Instant>) -> Self

pub fn termination_time(self, input: Instant) -> Self

A time stamp indicating when this data object was terminated. Format is a number expressed in Unix time as milliseconds (for example "1469498468.057").

pub fn set_termination_time(self, input: Option<Instant>) -> Self

pub fn current_player_session_count(self, input: i32) -> Self

Number of players currently in the game session.

pub fn set_current_player_session_count(self, input: Option<i32>) -> Self

pub fn maximum_player_session_count(self, input: i32) -> Self

The maximum number of players that can be connected simultaneously to the game session.

pub fn set_maximum_player_session_count(self, input: Option<i32>) -> Self

pub fn status(self, input: GameSessionStatus) -> Self

Current status of the game session. A game session must have an ACTIVE status to have player sessions.

pub fn set_status(self, input: Option<GameSessionStatus>) -> Self

pub fn status_reason(self, input: GameSessionStatusReason) -> Self

Provides additional information about game session status. INTERRUPTED indicates that the game session was hosted on a spot instance that was reclaimed, causing the active game session to be terminated.

pub fn set_status_reason(self, input: Option<GameSessionStatusReason>) -> Self

pub fn game_properties(self, input: impl Into<GameProperty>) -> Self

pub fn set_game_properties(self, input: Option<Vec<GameProperty>>) -> Self

pub fn ip_address(self, input: impl Into<String>) -> Self

The IP address of the game session. To connect to a GameLift game server, an app needs both the IP address and port number.

pub fn set_ip_address(self, input: Option<String>) -> Self

pub fn dns_name(self, input: impl Into<String>) -> Self

The DNS identifier assigned to the instance that is running the game session. Values have the following format:

TLS-enabled fleets: ..amazongamelift.com.

Non-TLS-enabled fleets: ec2-.compute.amazonaws.com. (See Amazon EC2 Instance IP Addressing.)

When connecting to a game session that is running on a TLS-enabled fleet, you must use the DNS name, not the IP address.

pub fn set_dns_name(self, input: Option<String>) -> Self

pub fn port(self, input: i32) -> Self

The port number for the game session. To connect to a GameLift game server, an app needs both the IP address and port number.

pub fn set_port(self, input: Option<i32>) -> Self

pub fn player_session_creation_policy(
self,
input: PlayerSessionCreationPolicy
) -> Self

Indicates whether or not the game session is accepting new players.

pub fn set_player_session_creation_policy(
self,
input: Option<PlayerSessionCreationPolicy>
) -> Self

pub fn creator_id(self, input: impl Into<String>) -> Self

A unique identifier for a player. This ID is used to enforce a resource protection policy (if one exists), that limits the number of game sessions a player can create.

pub fn set_creator_id(self, input: Option<String>) -> Self

pub fn game_session_data(self, input: impl Into<String>) -> Self

A set of custom game session properties, formatted as a single string value. This data is passed to a game server process in the GameSession object with a request to start a new game session.

pub fn set_game_session_data(self, input: Option<String>) -> Self

pub fn matchmaker_data(self, input: impl Into<String>) -> Self

Information about the matchmaking process that was used to create the game session. It is in JSON syntax, formatted as a string. In addition the matchmaking configuration used, it contains data on all players assigned to the match, including player attributes and team assignments. For more details on matchmaker data, see Match Data. Matchmaker data is useful when requesting match backfills, and is updated whenever new players are added during a successful backfill (see StartMatchBackfill).

pub fn set_matchmaker_data(self, input: Option<String>) -> Self

pub fn location(self, input: impl Into<String>) -> Self

The fleet location where the game session is running. This value might specify the fleet's home Region or a remote location. Location is expressed as an AWS Region code such as us-west-2.

pub fn set_location(self, input: Option<String>) -> Self

pub fn build(self) -> GameSession

Consumes the builder and constructs a GameSession