Struct aws_sdk_gamelift::model::FleetAttributes

#[non_exhaustive]
pub struct FleetAttributes {

    pub fleet_id: Option<String>,
    pub fleet_arn: Option<String>,
    pub fleet_type: Option<FleetType>,
    pub instance_type: Option<Ec2InstanceType>,
    pub description: Option<String>,
    pub name: Option<String>,
    pub creation_time: Option<Instant>,
    pub termination_time: Option<Instant>,
    pub status: Option<FleetStatus>,
    pub build_id: Option<String>,
    pub build_arn: Option<String>,
    pub script_id: Option<String>,
    pub script_arn: Option<String>,
    pub server_launch_path: Option<String>,
    pub server_launch_parameters: Option<String>,
    pub log_paths: Option<Vec<String>>,
    pub new_game_session_protection_policy: Option<ProtectionPolicy>,
    pub operating_system: Option<OperatingSystem>,
    pub resource_creation_limit_policy: Option<ResourceCreationLimitPolicy>,
    pub metric_groups: Option<Vec<String>>,
    pub stopped_actions: Option<Vec<FleetAction>>,
    pub instance_role_arn: Option<String>,
    pub certificate_configuration: Option<CertificateConfiguration>,

}

Describes a GameLift fleet of game hosting resources.

Related actions

CreateFleet | DescribeFleetAttributes

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

fleet_id: Option<String>

A unique identifier for the fleet.

fleet_arn: Option<String>

The Amazon Resource Name (ARN) that is assigned to a GameLift fleet resource and uniquely identifies it. ARNs are unique across all Regions. Format is arn:aws:gamelift:::fleet/fleet-a1234567-b8c9-0d1e-2fa3-b45c6d7e8912. In a GameLift fleet ARN, the resource ID matches the FleetId value.

fleet_type: Option<FleetType>

The kind of instances, On-Demand or Spot, that this fleet uses.

instance_type: Option<Ec2InstanceType>

The EC2 instance type that determines the computing resources of each instance in the fleet. Instance type defines the CPU, memory, storage, and networking capacity. See Amazon EC2 Instance Types for detailed descriptions.

description: Option<String>

A human-readable description of the fleet.

name: Option<String>

A descriptive label that is associated with a fleet. Fleet names do not need to be unique.

creation_time: Option<Instant>

A time stamp indicating when this data object was created. Format is a number expressed in Unix time as milliseconds (for example "1469498468.057").

termination_time: Option<Instant>

A time stamp indicating when this data object was terminated. Format is a number expressed in Unix time as milliseconds (for example "1469498468.057").

status: Option<FleetStatus>

Current status of the fleet. Possible fleet statuses include the following:

NEW -- A new fleet has been defined and desired instances is set to 1.

DOWNLOADING/VALIDATING/BUILDING/ACTIVATING -- GameLift is setting up the new fleet, creating new instances with the game build or Realtime script and starting server processes.

ACTIVE -- Hosts can now accept game sessions.

ERROR -- An error occurred when downloading, validating, building, or activating the fleet.

DELETING -- Hosts are responding to a delete fleet request.

TERMINATED -- The fleet no longer exists.

build_id: Option<String>

A unique identifier for the build resource that is deployed on instances in this fleet.

build_arn: Option<String>

The Amazon Resource Name (ARN) associated with the GameLift build resource that is deployed on instances in this fleet. In a GameLift build ARN, the resource ID matches the BuildId value.

script_id: Option<String>

A unique identifier for the Realtime script resource that is deployed on instances in this fleet.

script_arn: Option<String>

The Amazon Resource Name (ARN) associated with the GameLift script resource that is deployed on instances in this fleet. In a GameLift script ARN, the resource ID matches the ScriptId value.

server_launch_path: Option<String>

This parameter is no longer used. Server launch paths are now defined using the fleet's RuntimeConfiguration parameter. Requests that use this parameter instead continue to be valid.

server_launch_parameters: Option<String>

This parameter is no longer used. Server launch parameters are now defined using the fleet's RuntimeConfiguration parameter. Requests that use this parameter instead continue to be valid.

log_paths: Option<Vec<String>>

This parameter is no longer used. Game session log paths are now defined using the GameLift server API ProcessReady() logParameters. See more information in the Server API Reference.

new_game_session_protection_policy: Option<ProtectionPolicy>

The type of game session protection to set on all new instances that are started in the fleet.

NoProtection -- The game session can be terminated during a scale-down event.

FullProtection -- If the game session is in an ACTIVE status, it cannot be terminated during a scale-down event.

operating_system: Option<OperatingSystem>

The operating system of the fleet's computing resources. A fleet's operating system is determined by the OS of the build or script that is deployed on this fleet.

resource_creation_limit_policy: Option<ResourceCreationLimitPolicy>

The fleet policy that limits the number of game sessions an individual player can create over a span of time.

metric_groups: Option<Vec<String>>

Name of a metric group that metrics for this fleet are added to. In Amazon CloudWatch, you can view aggregated metrics for fleets that are in a metric group. A fleet can be included in only one metric group at a time.

stopped_actions: Option<Vec<FleetAction>>

A list of fleet activity that has been suspended using StopFleetActions. This includes fleet auto-scaling.

instance_role_arn: Option<String>

A unique identifier for an AWS IAM role that manages access to your AWS services. With an instance role ARN set, any application that runs on an instance in this fleet can assume the role, including install scripts, server processes, and daemons (background processes). Create a role or look up a role's ARN by using the IAM dashboard in the AWS Management Console. Learn more about using on-box credentials for your game servers at Access external resources from a game server.

certificate_configuration: Option<CertificateConfiguration>

Indicates whether a TLS/SSL certificate was generated for the fleet.

Implementations

impl FleetAttributes

pub fn builder() -> Builder

Creates a new builder-style object to manufacture FleetAttributes