Module aws_sdk_gamelift::operation

Structs

AcceptMatch

Registers a player's acceptance or rejection of a proposed FlexMatch match. A

ClaimGameServer

CreateAlias

Creates an alias for a fleet. In most situations, you can use an alias ID in place of

CreateBuild

Creates a new Amazon GameLift build resource for your game server binary files. Game server

CreateFleet

Creates a fleet of Amazon Elastic Compute Cloud (Amazon EC2) instances to host your

CreateFleetLocations

Adds remote locations to a fleet and begins populating the new locations with EC2

CreateGameServerGroup

CreateGameSession

Creates a multiplayer game session for players in a specific fleet location. This

CreateGameSessionQueue

Creates a placement queue that processes requests for new game sessions. A queue uses

CreateMatchmakingConfiguration

Defines a new matchmaking configuration for use with FlexMatch. Whether your are using

CreateMatchmakingRuleSet

Creates a new rule set for FlexMatch matchmaking. A rule set describes the type of match

CreatePlayerSession

Reserves an open player slot in a game session for a player. New player sessions can be

CreatePlayerSessions

Reserves open slots in a game session for a group of players. New player sessions can be

CreateScript

Creates a new script record for your Realtime Servers script. Realtime scripts are JavaScript that

CreateVpcPeeringAuthorization

Requests authorization to create or delete a peer connection between the VPC for

CreateVpcPeeringConnection

Establishes a VPC peering connection between a virtual private cloud (VPC) in an AWS account with the VPC

DeleteAlias

Deletes an alias. This operation removes all record of the alias. Game clients

DeleteBuild

Deletes a build. This operation permanently deletes the build resource and any uploaded

DeleteFleet

Deletes all resources and information related a fleet. Any current fleet instances,

DeleteFleetLocations

Removes locations from a multi-location fleet. When deleting a location, all game

DeleteGameServerGroup

DeleteGameSessionQueue

Deletes a game session queue. Once a queue is successfully deleted, unfulfilled StartGameSessionPlacement requests that reference the queue will fail.

DeleteMatchmakingConfiguration

Permanently removes a FlexMatch matchmaking configuration. To delete, specify the

DeleteMatchmakingRuleSet

Deletes an existing matchmaking rule set. To delete the rule set, provide the rule

DeleteScalingPolicy

Deletes a fleet scaling policy. Once deleted, the policy is no longer in

DeleteScript

Deletes a Realtime script. This operation permanently deletes the script record. If

DeleteVpcPeeringAuthorization

Cancels a pending VPC peering authorization for the specified VPC. If you need to

DeleteVpcPeeringConnection

Removes a VPC peering connection. To delete the connection, you must have a valid

DeregisterGameServer

DescribeAlias

Retrieves properties for an alias. This operation returns all alias metadata and

DescribeBuild

Retrieves properties for a custom game build. To request a build resource, specify a

DescribeEC2InstanceLimits

The GameLift service limits and current utilization for an AWS Region or location.

DescribeFleetAttributes

Retrieves core fleet-wide properties, including the computing hardware and deployment

DescribeFleetCapacity

Retrieves the resource capacity settings for one or more fleets. The data returned

DescribeFleetEvents

Retrieves entries from a fleet's event log. Fleet events are initiated by changes in

DescribeFleetLocationAttributes

Retrieves information on a fleet's remote locations, including life-cycle status and

DescribeFleetLocationCapacity

Retrieves the resource capacity settings for a fleet location. The data returned

DescribeFleetLocationUtilization

Retrieves current usage data for a fleet location. Utilization data provides a

DescribeFleetPortSettings

Retrieves a fleet's inbound connection permissions. Connection permissions specify the

DescribeFleetUtilization

Retrieves utilization statistics for one or more fleets. Utilization data provides a

DescribeGameServer

DescribeGameServerGroup

DescribeGameServerInstances

DescribeGameSessionDetails

Retrieves additional game session properties, including the game session protection

DescribeGameSessionPlacement

Retrieves information, including current status, about a game session placement request.

DescribeGameSessionQueues

Retrieves the properties for one or more game session queues. When requesting

DescribeGameSessions

Retrieves a set of one or more game sessions in a specific fleet location. You can

DescribeInstances

Retrieves information about a fleet's instances, including instance IDs, connection

DescribeMatchmaking

Retrieves one or more matchmaking tickets. Use this operation to retrieve ticket

DescribeMatchmakingConfigurations

Retrieves the details of FlexMatch matchmaking configurations.

DescribeMatchmakingRuleSets

Retrieves the details for FlexMatch matchmaking rule sets. You can request all

DescribePlayerSessions

Retrieves properties for one or more player sessions.

DescribeRuntimeConfiguration

Retrieves a fleet's runtime configuration settings. The runtime configuration tells

DescribeScalingPolicies

Retrieves all scaling policies applied to a fleet.

DescribeScript

Retrieves properties for a Realtime script.

DescribeVpcPeeringAuthorizations

Retrieves valid VPC peering authorizations that are pending for the AWS account.

DescribeVpcPeeringConnections

Retrieves information on VPC peering connections. Use this operation to get peering

GetGameSessionLogUrl

Retrieves the location of stored game session logs for a specified game session.

GetInstanceAccess

Requests remote access to a fleet instance. Remote access is useful for debugging,

ListAliases

Retrieves all aliases for this AWS account. You can filter the result set by

ListBuilds

Retrieves build resources for all builds associated with the AWS account in use. You

ListFleets

Retrieves a collection of fleet resources in an AWS Region. You can call this operation

ListGameServerGroups

ListGameServers

ListScripts

Retrieves script records for all Realtime scripts that are associated with the AWS account in use.

ListTagsForResource

PutScalingPolicy

Creates or updates a scaling policy for a fleet. Scaling policies are used to

RegisterGameServer

RequestUploadCredentials

Retrieves a fresh set of credentials for use when uploading a new set of game build

ResolveAlias

Retrieves the fleet ID that an alias is currently pointing to.

ResumeGameServerGroup

SearchGameSessions

Retrieves all active game sessions that match a set of search criteria and sorts them

StartFleetActions

Resumes certain types of activity on fleet instances that were suspended with StopFleetActions. For multi-location fleets, fleet actions are managed

StartGameSessionPlacement

Places a request for a new game session in a queue (see CreateGameSessionQueue). When processing a placement request, Amazon GameLift

StartMatchBackfill

Finds new players to fill open slots in currently running game sessions.

StartMatchmaking

Uses FlexMatch to create a game match for a group of players based on custom matchmaking

StopFleetActions

Suspends certain types of activity in a fleet location. Currently, this operation is

StopGameSessionPlacement

Cancels a game session placement that is in PENDING status. To stop a

StopMatchmaking

Cancels a matchmaking ticket or match backfill ticket that is currently being

SuspendGameServerGroup

TagResource

Assigns a tag to a GameLift resource. AWS resource tags provide an additional

UntagResource

Removes a tag that is assigned to a GameLift resource.

UpdateAlias

Updates properties for an alias. To update properties, specify the alias ID to be

UpdateBuild

Updates metadata in a build resource, including the build name and version. To update

UpdateFleetAttributes

Updates a fleet's mutable attributes, including game session protection and resource

UpdateFleetCapacity

Updates capacity settings for a fleet. For fleets with multiple locations, use this

UpdateFleetPortSettings

Updates permissions that allow inbound traffic to connect to game sessions that are

UpdateGameServer

UpdateGameServerGroup

UpdateGameSession

Updates the mutable properties of a game session.

UpdateGameSessionQueue

Updates the configuration of a game session queue, which determines how the queue

UpdateMatchmakingConfiguration

Updates settings for a FlexMatch matchmaking configuration. These changes affect all matches and game sessions

UpdateRuntimeConfiguration

Updates the current runtime configuration for the specified fleet, which tells GameLift

UpdateScript

Updates Realtime script metadata and content.

ValidateMatchmakingRuleSet

Validates the syntax of a matchmaking rule or rule set. This operation checks that