Module aws_sdk_gamelift::output

Modules

accept_match_output

See AcceptMatchOutput

claim_game_server_output

See ClaimGameServerOutput

create_alias_output

See CreateAliasOutput

create_build_output

See CreateBuildOutput

create_fleet_locations_output

See CreateFleetLocationsOutput

create_fleet_output

See CreateFleetOutput

create_game_server_group_output

See CreateGameServerGroupOutput

create_game_session_output

See CreateGameSessionOutput

create_game_session_queue_output

See CreateGameSessionQueueOutput

create_matchmaking_configuration_output

See CreateMatchmakingConfigurationOutput

create_matchmaking_rule_set_output

See CreateMatchmakingRuleSetOutput

create_player_session_output

See CreatePlayerSessionOutput

create_player_sessions_output

See CreatePlayerSessionsOutput

create_script_output

See CreateScriptOutput

create_vpc_peering_authorization_output

See CreateVpcPeeringAuthorizationOutput

create_vpc_peering_connection_output

See CreateVpcPeeringConnectionOutput

delete_alias_output

See DeleteAliasOutput

delete_build_output

See DeleteBuildOutput

delete_fleet_locations_output

See DeleteFleetLocationsOutput

delete_fleet_output

See DeleteFleetOutput

delete_game_server_group_output

See DeleteGameServerGroupOutput

delete_game_session_queue_output

See DeleteGameSessionQueueOutput

delete_matchmaking_configuration_output

See DeleteMatchmakingConfigurationOutput

delete_matchmaking_rule_set_output

See DeleteMatchmakingRuleSetOutput

delete_scaling_policy_output

See DeleteScalingPolicyOutput

delete_script_output

See DeleteScriptOutput

delete_vpc_peering_authorization_output

See DeleteVpcPeeringAuthorizationOutput

delete_vpc_peering_connection_output

See DeleteVpcPeeringConnectionOutput

deregister_game_server_output

See DeregisterGameServerOutput

describe_alias_output

See DescribeAliasOutput

describe_build_output

See DescribeBuildOutput

describe_ec2_instance_limits_output

See DescribeEc2InstanceLimitsOutput

describe_fleet_attributes_output

See DescribeFleetAttributesOutput

describe_fleet_capacity_output

See DescribeFleetCapacityOutput

describe_fleet_events_output

See DescribeFleetEventsOutput

describe_fleet_location_attributes_output

See DescribeFleetLocationAttributesOutput

describe_fleet_location_capacity_output

See DescribeFleetLocationCapacityOutput

describe_fleet_location_utilization_output

See DescribeFleetLocationUtilizationOutput

describe_fleet_port_settings_output

See DescribeFleetPortSettingsOutput

describe_fleet_utilization_output

See DescribeFleetUtilizationOutput

describe_game_server_group_output

See DescribeGameServerGroupOutput

describe_game_server_instances_output

See DescribeGameServerInstancesOutput

describe_game_server_output

See DescribeGameServerOutput

describe_game_session_details_output

See DescribeGameSessionDetailsOutput

describe_game_session_placement_output

See DescribeGameSessionPlacementOutput

describe_game_session_queues_output

See DescribeGameSessionQueuesOutput

describe_game_sessions_output

See DescribeGameSessionsOutput

describe_instances_output

See DescribeInstancesOutput

describe_matchmaking_configurations_output

See DescribeMatchmakingConfigurationsOutput

describe_matchmaking_output

See DescribeMatchmakingOutput

describe_matchmaking_rule_sets_output

See DescribeMatchmakingRuleSetsOutput

describe_player_sessions_output

See DescribePlayerSessionsOutput

describe_runtime_configuration_output

See DescribeRuntimeConfigurationOutput

describe_scaling_policies_output

See DescribeScalingPoliciesOutput

describe_script_output

See DescribeScriptOutput

describe_vpc_peering_authorizations_output

See DescribeVpcPeeringAuthorizationsOutput

describe_vpc_peering_connections_output

See DescribeVpcPeeringConnectionsOutput

get_game_session_log_url_output

See GetGameSessionLogUrlOutput

get_instance_access_output

See GetInstanceAccessOutput

list_aliases_output

See ListAliasesOutput

list_builds_output

See ListBuildsOutput

list_fleets_output

See ListFleetsOutput

list_game_server_groups_output

See ListGameServerGroupsOutput

list_game_servers_output

See ListGameServersOutput

list_scripts_output

See ListScriptsOutput

list_tags_for_resource_output

See ListTagsForResourceOutput

put_scaling_policy_output

See PutScalingPolicyOutput

register_game_server_output

See RegisterGameServerOutput

request_upload_credentials_output

See RequestUploadCredentialsOutput

resolve_alias_output

See ResolveAliasOutput

resume_game_server_group_output

See ResumeGameServerGroupOutput

search_game_sessions_output

See SearchGameSessionsOutput

start_fleet_actions_output

See StartFleetActionsOutput

start_game_session_placement_output

See StartGameSessionPlacementOutput

start_match_backfill_output

See StartMatchBackfillOutput

start_matchmaking_output

See StartMatchmakingOutput

stop_fleet_actions_output

See StopFleetActionsOutput

stop_game_session_placement_output

See StopGameSessionPlacementOutput

stop_matchmaking_output

See StopMatchmakingOutput

suspend_game_server_group_output

See SuspendGameServerGroupOutput

tag_resource_output

See TagResourceOutput

untag_resource_output

See UntagResourceOutput

update_alias_output

See UpdateAliasOutput

update_build_output

See UpdateBuildOutput

update_fleet_attributes_output

See UpdateFleetAttributesOutput

update_fleet_capacity_output

See UpdateFleetCapacityOutput

update_fleet_port_settings_output

See UpdateFleetPortSettingsOutput

update_game_server_group_output

See UpdateGameServerGroupOutput

update_game_server_output

See UpdateGameServerOutput

update_game_session_output

See UpdateGameSessionOutput

update_game_session_queue_output

See UpdateGameSessionQueueOutput

update_matchmaking_configuration_output

See UpdateMatchmakingConfigurationOutput

update_runtime_configuration_output

See UpdateRuntimeConfigurationOutput

update_script_output

See UpdateScriptOutput

validate_matchmaking_rule_set_output

See ValidateMatchmakingRuleSetOutput

Structs

AcceptMatchOutput

ClaimGameServerOutput

CreateAliasOutput

Represents the returned data in response to a request operation.

CreateBuildOutput

Represents the returned data in response to a request operation.

CreateFleetLocationsOutput

Represents the returned data in response to a request operation.

CreateFleetOutput

Represents the returned data in response to a request operation.

CreateGameServerGroupOutput

CreateGameSessionOutput

Represents the returned data in response to a request operation.

CreateGameSessionQueueOutput

Represents the returned data in response to a request operation.

CreateMatchmakingConfigurationOutput

Represents the returned data in response to a request operation.

CreateMatchmakingRuleSetOutput

Represents the returned data in response to a request operation.

CreatePlayerSessionOutput

Represents the returned data in response to a request operation.

CreatePlayerSessionsOutput

Represents the returned data in response to a request operation.

CreateScriptOutput

CreateVpcPeeringAuthorizationOutput

Represents the returned data in response to a request operation.

CreateVpcPeeringConnectionOutput

DeleteAliasOutput

DeleteBuildOutput

DeleteFleetLocationsOutput

Represents the returned data in response to a request operation.

DeleteFleetOutput

DeleteGameServerGroupOutput

DeleteGameSessionQueueOutput

DeleteMatchmakingConfigurationOutput

DeleteMatchmakingRuleSetOutput

Represents the returned data in response to a request operation.

DeleteScalingPolicyOutput

DeleteScriptOutput

DeleteVpcPeeringAuthorizationOutput

DeleteVpcPeeringConnectionOutput

DeregisterGameServerOutput

DescribeAliasOutput

Represents the returned data in response to a request operation.

DescribeBuildOutput

Represents the returned data in response to a request operation.

DescribeEc2InstanceLimitsOutput

Represents the returned data in response to a request operation.

DescribeFleetAttributesOutput

Represents the returned data in response to a request operation.

DescribeFleetCapacityOutput

Represents the returned data in response to a request operation.

DescribeFleetEventsOutput

Represents the returned data in response to a request operation.

DescribeFleetLocationAttributesOutput

Represents the returned data in response to a request operation.

DescribeFleetLocationCapacityOutput

Represents the returned data in response to a request operation.

DescribeFleetLocationUtilizationOutput

Represents the returned data in response to a request operation.

DescribeFleetPortSettingsOutput

Represents the returned data in response to a request operation.

DescribeFleetUtilizationOutput

Represents the returned data in response to a request operation.

DescribeGameServerGroupOutput

DescribeGameServerInstancesOutput

DescribeGameServerOutput

DescribeGameSessionDetailsOutput

Represents the returned data in response to a request operation.

DescribeGameSessionPlacementOutput

Represents the returned data in response to a request operation.

DescribeGameSessionQueuesOutput

Represents the returned data in response to a request operation.

DescribeGameSessionsOutput

Represents the returned data in response to a request operation.

DescribeInstancesOutput

Represents the returned data in response to a request operation.

DescribeMatchmakingConfigurationsOutput

Represents the returned data in response to a request operation.

DescribeMatchmakingOutput

Represents the returned data in response to a request operation.

DescribeMatchmakingRuleSetsOutput

Represents the returned data in response to a request operation.

DescribePlayerSessionsOutput

Represents the returned data in response to a request operation.

DescribeRuntimeConfigurationOutput

Represents the returned data in response to a request operation.

DescribeScalingPoliciesOutput

Represents the returned data in response to a request operation.

DescribeScriptOutput

DescribeVpcPeeringAuthorizationsOutput

DescribeVpcPeeringConnectionsOutput

Represents the returned data in response to a request operation.

GetGameSessionLogUrlOutput

Represents the returned data in response to a request operation.

GetInstanceAccessOutput

Represents the returned data in response to a request operation.

ListAliasesOutput

Represents the returned data in response to a request operation.

ListBuildsOutput

Represents the returned data in response to a request operation.

ListFleetsOutput

Represents the returned data in response to a request operation.

ListGameServerGroupsOutput

ListGameServersOutput

ListScriptsOutput

ListTagsForResourceOutput

PutScalingPolicyOutput

Represents the returned data in response to a request operation.

RegisterGameServerOutput

RequestUploadCredentialsOutput

Represents the returned data in response to a request operation.

ResolveAliasOutput

Represents the returned data in response to a request operation.

ResumeGameServerGroupOutput

SearchGameSessionsOutput

Represents the returned data in response to a request operation.

StartFleetActionsOutput

Represents the returned data in response to a request operation.

StartGameSessionPlacementOutput

Represents the returned data in response to a request operation.

StartMatchBackfillOutput

Represents the returned data in response to a request operation.

StartMatchmakingOutput

Represents the returned data in response to a request operation.

StopFleetActionsOutput

Represents the input for a request operation.

StopGameSessionPlacementOutput

Represents the returned data in response to a request operation.

StopMatchmakingOutput

SuspendGameServerGroupOutput

TagResourceOutput

UntagResourceOutput

UpdateAliasOutput

Represents the returned data in response to a request operation.

UpdateBuildOutput

Represents the returned data in response to a request operation.

UpdateFleetAttributesOutput

Represents the returned data in response to a request operation.

UpdateFleetCapacityOutput

Represents the returned data in response to a request operation.

UpdateFleetPortSettingsOutput

Represents the returned data in response to a request operation.

UpdateGameServerGroupOutput

UpdateGameServerOutput

UpdateGameSessionOutput

Represents the returned data in response to a request operation.

UpdateGameSessionQueueOutput

Represents the returned data in response to a request operation.

UpdateMatchmakingConfigurationOutput

Represents the returned data in response to a request operation.

UpdateRuntimeConfigurationOutput

Represents the returned data in response to a request operation.

UpdateScriptOutput

ValidateMatchmakingRuleSetOutput

Represents the returned data in response to a request operation.